requirements.txt
pipecat-ai[deepgram,openai,cartesia,silero,simli,webrtc]
fastapi
uvicorn
aiohttp
python-dotenv
loguru